% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minerr.R
\name{minerr_log}
\alias{minerr_log}
\title{Either print or download the last log of the communication with AI}
\usage{
minerr_log(download = FALSE)
}
\arguments{
\item{download}{logical, if set to TRUE, it'll download the log file as .txt}
}
\value{
character or a file, the log of the chat with AI
}
\description{
Either print or download the last log of the communication with AI
}
\examples{
\dontrun{
minerr_log()
minerr_log(download = TRUE)
}
}
