% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minerr.R
\name{minerr}
\alias{minerr}
\title{Using AI/LLM analyzes the given error and offers solution}
\usage{
minerr(
  paren.frame = parent.frame(),
  chat = NULL,
  share_full_context = TRUE,
  llm_func = purrr::partial(ellmer::chat_openai, model = "gpt-4o"),
  insert_code = TRUE
)
}
\arguments{
\item{paren.frame}{environment, parent context. Not needed to provide any value.}

\item{chat}{character, user input}

\item{share_full_context}{logical, whether to share the context or not}

\item{llm_func}{function, preferred chat model. default it openai.}

\item{insert_code}{logical, whether to insert code in the editor or not}
}
\value{
LLM's response, both in the console and inserts into the editor
}
\description{
Using AI/LLM analyzes the given error and offers solution
}
\examples{
\dontrun{
library(data.table)
dt <- as.data.table(mtcars)
dt[, .(cnt = .N), by = .("cyl", "gear")] |> minerr()
}
}
